refactor: clean, reorganize
