+++ SwiftDiffusion/App/AppStructure/App+Setup.swift
