+++ SwiftDiffusion/App/AppStructure/AppFileStructure.swift
