+++ SwiftDiffusion/App/Constants.swift
