+++ SwiftDiffusion/ContentView/ContentView+Generate.swift
