+++ SwiftDiffusion/DetailView/DetailView.swift
