+++ SwiftDiffusion/DetailView/Extensions/NSImageExtensions.swift
