+++ SwiftDiffusion/DetailView/FileHierarchy/FileNode.swift
