+++ SwiftDiffusion/DetailView/FileOutlineView.swift
