+++ SwiftDiffusion/DetailView/FileRowView.swift
