+++ SwiftDiffusion/DetailView/ImageDetailViews/FullscreenImageView.swift
