+++ SwiftDiffusion/DetailView/ImageDetailViews/ImageWindowManager.swift
