+++ SwiftDiffusion/DetailView/ProgressView.swift
