+++ SwiftDiffusion/DetailView/ThumbnailLoader.swift
