+++ SwiftDiffusion/PromptView/DebugPromptViews/DebugPromptStatusView.swift
