+++ SwiftDiffusion/PromptView/PromptModels/PromptModel.swift
