+++ SwiftDiffusion/PromptView/PromptModels/SdModelData.swift
