+++ SwiftDiffusion/PromptView/PromptStatusBarViews/PromptBottomStatusBar.swift
