+++ SwiftDiffusion/PromptView/PromptStatusBarViews/PromptTopStatusBar.swift
