+++ SwiftDiffusion/PromptView/PromptView+ParseCivitai.swift
