+++ SwiftDiffusion/SettingsView/RequiredInputPaths/RequiredInputPathsPulsatingButton.swift
