+++ SwiftDiffusion/Views/UpdatesView/UpdatesView.swift
